% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellany.R
\name{convert_decimal_time}
\alias{convert_decimal_time}
\title{A Function to Convert Decimal Time to Days, Hours, Minutes}
\usage{
convert_decimal_time(days = 0, hours = 0)
}
\arguments{
\item{days}{Number of days, must be numeric and non-negative}

\item{hours}{Number of hours, must be numeric and non-negative}
}
\value{
A named list of day, hour, and minute amounts for the decimal time
        given. Will also print this to the console.
}
\description{
This function allows for easy conversion of time give as decimals to days,
minutes, and hours. Useful for switching between different modes of tracking
time.
}
\examples{
convert_decimal_time(days = 3, hours = 2.5)
"Total time is 3 days, 2 hours, and 30 minutes."

convert_decimal_time(days = 1.5)
"Total time is 1 days, 3 hours, and 42 minutes."

convert_decimal_time(hours = 6.25)
"Total time is 0 days, 6 hours, and 15 minutes."
}
