% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Compute Geometric Mean of a Dataset}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{An atomic vector of numeric type.}
}
\description{
This will compute the GM for a set of data
}
\examples{
gm_mean(1:10)
[1] 4.528729
}
