% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{hm_mean}
\alias{hm_mean}
\title{Compute Harmonic Mean of a Dataset}
\usage{
hm_mean(x)
}
\arguments{
\item{x}{An atomic vector of numeric type.}
}
\description{
This will compute the HM for a set of data
}
\examples{
hm_mean(1:10)
3.414172
}
