% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_mgmt.R
\name{install_new_pkgs}
\alias{install_new_pkgs}
\title{Keeping Installed Packages Consistent Across Environments}
\usage{
install_new_pkgs(pkg_file = "HBH_R_pkgs.csv")
}
\arguments{
\item{pkg_file}{The file created from installed.packages() with default path.}
}
\description{
Useful function that can read in any csv output from installed.packages() and
attempt twice to install from CRAN any packages not locally installed (as
often the first attempt will fail because of missed dependencies). It will
then source from Bioconductor and attempt to install the remaining packages
from there, finally returning a character vector of the packages that could
not be installed from either of those sources. These will often need to
be installed using devtools::install_github().
}
