% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_mgmt.R
\name{write_R_pkg_list}
\alias{write_R_pkg_list}
\title{Update Saved List of Installed Packages}
\usage{
write_R_pkg_list(tools_dir = NULL)
}
\arguments{
\item{tools_dir}{Defaults to NULL, can pass a directory explicitly if needed}
}
\description{
Function to merge the list of locally installed packages with the saved list
of packages.
}
